src/pytomtom.py: Improve path management
Application and config paths are global in the scope
of the pytomtom application, so there is no need to
store them in the class context.

Move them outside the class and make them global.

Fix as well the pix and i18n path to make them relative
to the pytomtom.py location. It allows to run the
application from any places.

Signed-off-by: Benoit Cousson <benoit.cousson@gmail.com>